Add editor options collection builder. Move all setters to Field class.
