+++ src/Html/Editor/Select.php
@@ -5,17 +5,4 @@ namespace Yajra\DataTables\Html\Editor;
 class Select extends Field
 {
     protected $type = 'select';
-
-    /**
-     * Set select options.
-     *
-     * @param array $options
-     * @return $this
-     */
-    public function options(array $options)
-    {
-        $this->attributes['options'] = $options;
-
-        return $this;
-    }
 }
